% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglangparallel.R
\name{gglangparallel}
\alias{gglangparallel}
\title{Create a ggtextparallel by versions}
\usage{
gglangparallel(versions, book, verses, words_per_row = 5)
}
\arguments{
\item{versions}{valid versions obtained from biblesearch_versions}

\item{book}{book, e.g. "genesis", "mark", "romans"}

\item{words_per_row}{words displayed per rown}

\item{passage}{text, e.g. "1:1-5", "1.1-5"}
}
\value{
a ggplot2 object
}
\description{
Create a ggtextparallel by versions
}
