% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtextparallel.R
\name{ggtextparallel}
\alias{ggtextparallel}
\title{Create a ggtextparallel}
\usage{
ggtextparallel(parallel_no, lang = "eng")
}
\arguments{
\item{parallel_no}{integer from the No. column in gospel_parallels}

\item{lang}{language, "eng" for English, "grk" for Greek}
}
\value{
`ggplot2` object that facets the text pericopes
}
\description{
Create a ggtextparallel
}
\examples{
ggtextparallel(93)
}
