% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gospel_parallels}
\alias{gospel_parallels}
\title{Gospel paralells from Synopsis Quattuor Evangeliorum}
\format{A data frame with 367 rows and 6 variables:
\describe{
  \item{No.}{parallel number}
  \item{Pericope}{pericope title}
  \item{Matthew}{text citation from Matthew}
  \item{Mark}{text citation from Mark}
  \item{Luke}{text citation from Luke}
  \item{John}{text citation from John}
  ...
}}
\source{
\url{http://www.bible-researcher.com/parallels.html}
}
\usage{
gospel_parallels
}
\description{
A dataset containing gospel parallels
gospel_parallels.
}
\keyword{datasets}
