% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtextparallel.R
\name{ggtextparallel}
\alias{ggtextparallel}
\title{Create a ggtextparallel}
\usage{
ggtextparallel(parallel_no, version_id = NULL, words_per_row = 7)
}
\arguments{
\item{parallel_no}{integer from the No. column in gospel_parallels.}

\item{version_id}{translation version from versions object. "grc" for koine greek.}

\item{words_per_row}{how many words are displayed per row.}
}
\value{
`ggplot2` object that facets the text pericopes
}
\description{
Create a ggtextparallel
}
\examples{
ggtextparallel(93, version_id = "eng-ESV")
ggtextparallel(17, version_id = "grc", words_per_row = 5)
}
