% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtextparallel.R, R/view_pericope.R
\name{ggtextparallel}
\alias{ggtextparallel}
\alias{ggtextparallel}
\title{Create a ggtextparallel}
\usage{
ggtextparallel(pericope_no)

ggtextparallel(pericope_no)
}
\arguments{
\item{pericope_no}{integer from the No. column in gospel_parallels}

\item{parallel_no}{integer from the No. column in gospel_parallels}
}
\value{
`ggplot2` object that facets the text pericopes

`ggplot2` object that facets the text pericopes
}
\description{
Create a ggtextparallel

Create a ggtextparallel
}
\examples{
ggtextparallel(93)
ggtextparallel(93)
ggtextparallel(17)
}
